Make LazyHandle a top-level class.
