+++ core/src/main/java/org/jdbi/v3/DBI.java
@@ -25,7 +25,6 @@ import java.util.concurrent.ConcurrentHashMap;
 import java.util.concurrent.CopyOnWriteArrayList;
 import java.util.concurrent.atomic.AtomicReference;
 import java.util.function.Consumer;
-import java.util.function.Supplier;
 
 import javax.sql.DataSource;
 
@@ -398,7 +397,7 @@ public class DBI
     public <R, E, X extends Exception> R withExtension(Class<E> extensionType, ExtensionCallback<R, E, X> callback)
             throws NoSuchExtensionException, X
     {
-        try (LazyHandle handle = new LazyHandle()) {
+        try (LazyHandle handle = new LazyHandle(this)) {
             E extension = extensionRegistry.findExtensionFor(extensionType, handle)
                     .orElseThrow(() -> new NoSuchExtensionException("Extension not found: " + extensionType));
 
@@ -423,34 +422,6 @@ public class DBI
             callback.useExtension(extension);
             return null;
         });
-    }
-
-    private class LazyHandle implements Supplier<Handle>, AutoCloseable {
-        private volatile Handle handle;
-        private volatile boolean closed = false;
-
-        public Handle get() {
-            if (handle == null) {
-                initHandle();
-            }
-            return handle;
-        }
-
-        private synchronized void initHandle() {
-            if (handle == null) {
-                if (closed) {
-                    throw new IllegalStateException("Handle is closed");
-                }
-                handle = open();
-            }
-        }
-
-        public synchronized void close() {
-            closed = true;
-            if (handle != null) {
-                handle.close();
-            }
-        }
     }
 
     /**
